JBTM-3722 jakarta native migration
