--- javaee/src/main/java/org/jboss/tm/ConnectableLastResource.java
@@ -1,43 +1,0 @@
-/*
-  * JBoss, Home of Professional Open Source
-  * Copyright 2011, JBoss Inc., and individual contributors as indicated
-  * by the @authors tag. See the copyright.txt in the distribution for a
-  * full listing of individual contributors.
-  *
-  * This is free software; you can redistribute it and/or modify it
-  * under the terms of the GNU Lesser General Public License as
-  * published by the Free Software Foundation; either version 2.1 of
-  * the License, or (at your option) any later version.
-  *
-  * This software is distributed in the hope that it will be useful,
-  * but WITHOUT ANY WARRANTY; without even the implied warranty of
-  * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU
-  * Lesser General Public License for more details.
-  *
-  * You should have received a copy of the GNU Lesser General Public
-  * License along with this software; if not, write to the Free
-  * Software Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA
-  * 02110-1301 USA, or see the FSF site: http://www.fsf.org.
-  */
-package org.jboss.tm;
-
-/**
- * @deprecated use {@link ConnectableResource} and {@link LastResource} instead
- *
- * A connectable last resource
- * 
- * @author <a href="mailto:jesper.pedersen@jboss.org">Jesper Pedersen</a>
- * @author Jonathan Halliday (jonathan.halliday@redhat.com)
- */
-@Deprecated
-public interface ConnectableLastResource extends LastResource
-{
-   /**
-    * Get connection.
-    *
-    * @return The connection; otherwise <code>null</code> if the method isn't supported
-    * @exception Throwable If error occurs
-    */
-   public Object getConnection() throws Throwable;
-}
-
