+++ src/main/java/org/jboss/tm/ConnectableResource.java
