+++ src/main/java/org/jboss/tm/LastResource.java
