+++ src/main/java/org/jboss/tm/XAResourceWrapper.java
